requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
SQLAlchemy==2.0.35
Werkzeug==3.0.4
pytest==8.3.3
